% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_case}
\alias{mids_case}
\title{mids_case}
\format{
A list containing simplified mids for the case study
}
\usage{
data(mids_case)
}
\description{
A list of (simplified) mids objects and condition descriptions used to obtain the trace plots for desired conditions.
}
\keyword{datasets}
