% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids_sim}
\alias{mids_sim}
\title{mids_sim}
\format{
A list containing simplified mids for the simulation study
}
\usage{
data(mids_sim)
}
\description{
A list of (simplified) mids objects and condition descriptions used to obtain the trace plots for desired conditions.
}
\keyword{datasets}
