% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_case.R
\name{plot_case}
\alias{plot_case}
\title{Plot case study results}
\usage{
plot_case(results, y)
}
\arguments{
\item{results}{object containing results produced by the simulation study}

\item{y}{dependent variable in the substantive analysis model}
}
\value{
Returns the ggplot
}
\description{
Generate the main plot for the case study.
}
\examples{
# Define example inputs
results <- dataFdd
y <- "PTSD-RI parent score"

# Use the function
plot_case(
    results = dataFdd,
    y = "yp"
)

}
\author{
Edoardo Costantini, 2023
}
