% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plots}
\usage{
plot_trace(mids_data, method, iters = c(0, 25), layout = c(2, 1))
}
\arguments{
\item{mids_data}{object containing imputation history produced by the simulation study}

\item{method}{unit character vector naming the method to display}

\item{iters}{numeric vector containing the iteration bounds (from-to) to plot}

\item{var}{unit character vector naming the variable to display}
}
\value{
Returns the lattice plot
}
\description{
Generate trace plots to study convergence of imputation methods
}
\examples{
# Define example inputs
mids_data <- mids_case
method <- "expert"
layout <- c(2, 3)
iters <- c(0, 25)

# Use the function
plot_trace(
    mids_data = mids_sim,
    method = "MIMI",
    layout <- c(2, 4),
    iters = c(0, 100)
)

}
\author{
Edoardo Costantini, 2023
}
