% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plotMids}
\alias{plotMids}
\title{plotMids}
\usage{
plotMids(study = c("sim", "fdd")[1])
}
\arguments{
\item{study.}{A unit character vector taking value "sim" or "fdd" (simulation study and case study, respectively)}
}
\value{
Shiny app UI.
}
\description{
Starts a Shiny app to check trace plots for multiple imputation convergence for the \href{https://github.com/EdoardoCostantini/mi-spcr}{mi-spcr} project.
}
\details{
The interface of the Shiny app allows you to change the values of the following simulation study experimental factors:
\itemize{
\item Missing data treatment used (see names in the interface):.cols
\itemize{
\item for the simulation study:
\itemize{
\item MIMI: mice with minimal missing data models (only variables under imputation used as predictors in the imputation models)
\item MIOP: mice using the quickpred function from the R package mice (Van Buuren and Groothuis-Oudshoorn, 2011) to select the predictors for the univariate imputation models via the correlation-based threshold
\item MIOR: mice with oracle knowledge on which predictors to use (the univariate imputation models included the other variables under imputation and the predictors that were used to impose missingness)
\item aux: mi-pca using pca only on auxiliary variables
\item vbv: mi-pca using pca only on a variable-by-variable basis
\item all: mi-pca using pca on all data, after a round of single imputation
}
\item for the case study:
\itemize{
\item expert: mice specified by an expert
\item si4auxall: convergence check for single imputation used for pre-processing for mi-pca-all
\item pcraux: mi-pca using pca only on auxiliary variables
\item vbv: mi-pca using pca only on a variable-by-variable basis
\item default: mice specified with default arguments
}
}
\item Imputed variable
\item Number of iterations used
}
}
