% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plotmipca}
\alias{plotmipca}
\title{plotmipca}
\usage{
plotmipca(...)
}
\value{
None
}
\description{
Starts a Shiny app to interact with the results of the \href{https://github.com/EdoardoCostantini/mi-pca}{mi-pca}.
}
\examples{
plotmipca()
}
