% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plotResults}
\alias{plotResults}
\title{plotResults}
\usage{
plotResults()
}
\value{
Shiny app UI.
}
\description{
Starts a Shiny app to interact with the results of the \href{https://github.com/EdoardoCostantini/mi-pca}{mi-pca} project.
}
