% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_plots.R
\name{trace_plot}
\alias{trace_plot}
\title{Trace plots}
\usage{
trace_plot(mids_data, method, var, iters = c(0, 25))
}
\arguments{
\item{mids_data}{object containing imputation history produced by the simulation study}

\item{method}{unit character vector naming the method to display}

\item{var}{unit character vector naming the variable to display}

\item{iters}{numeric vector containing the iteration bounds (from-to) to plot}
}
\value{
Returns the lattice plot
}
\description{
Generate trace plots to study convergence of imputation methods
}
\examples{
# Define example inputs
mids_data <- dataMids$sim
method <- "MIMI"
var <- "z1"
iters <- c(0, 25)

# Use the function
trace_plot(
    mids_data = dataMids$sim,
    method = "MIMI",
    var = "z1",
    iters = c(0, 100)
)

}
\author{
Edoardo Costantini, 2023
}
