% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataMids}
\alias{dataMids}
\title{dataMids}
\format{
A list containing (1) simplified mids for 81 conditions described in a (2) data.frame
}
\usage{
data(dataMids)
}
\description{
A list of (simplified) mids objects and condition descriptions used to obtain the trace plots for desired conditions.
}
\keyword{datasets}
